+++ src/domain/gamification/index.ts
@@ -1,0 +1,42 @@
+export {
+  levelFor,
+  type LevelProgress,
+  levelProgress,
+  RANK_LADDER,
+  rankFor,
+  xpToReach,
+} from './level';
+export {
+  type Crowns,
+  crownsFor,
+  domainMastery,
+  MAX_CROWNS,
+  type NextCrownHint,
+  nextCrownHint,
+  NO_PROGRESS,
+  type ObjectiveProgress,
+} from './mastery';
+export {
+  DAYS_PER_FREEZE,
+  type DayReport,
+  INITIAL_STREAK,
+  isStreakAlive,
+  MAX_FREEZES,
+  recordDay,
+  type StreakOutcome,
+  type StreakState,
+  streakDeadline,
+} from './streak';
+export {
+  addToLedger,
+  award,
+  type DailyLedger,
+  examXp,
+  pruneLedger,
+  quizXp,
+  totalOf,
+  XP,
+  type XpAward,
+  xpOn,
+  type XpReason,
+} from './xp';